requirements.txt
opencv-python
pyyaml
json_tricks
yacs
scikit-learn
pandas
timm==0.4.12
numpy==1.23.5
detectron2 @ git+https://github.com/facebookresearch/detectron2.git#864913f0e57e87a75c8cc0c7d79ecbd774fc669b
einops
fvcore
transformers==4.19.2
sentencepiece
ftfy
regex
nltk
vision-datasets==0.2.2
pycocotools==2.0.7
diffdist
pyarrow
cityscapesscripts
shapely
scikit-image
mup
gradio==3.13.0
scann
kornia==0.6.4
torchmetrics==0.6.0
mpi4py